Rename domain error to errors
