+++ apps/api/src/domain/errors.py
