+++ apps/api/src/infrastructure/handlers/streetlights_list.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from functools import lru_cache
 
-from domain.error import AuthError
+from domain.errors import AuthError
 from application.streetlight.list_streetlights import ListStreetlights
 from application.streetlight.responses import streetlight_to_list_item
 from infrastructure.persistence.dynamo.streetlights_repo import (
