+++ apps/api/src/infrastructure/handlers/websocket_connect.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 from functools import lru_cache
 
 from application.websocket.connect import ConnectWebSocket
-from domain.error import AuthError
+from domain.errors import AuthError
 from infrastructure.auth.identity import IdentityResolver
 from infrastructure.persistence.dynamo.websocket_connection_repo import (
     get_websocket_connection_repo
